Add b&b dfs part 1

+++ ReasearchPlatform/Models/Distance.cs
@@ -1,5 +1,4 @@
 ﻿using ResearchPlatform.Models.DTO;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -9,8 +8,8 @@ namespace ResearchPlatform.Models
     {
         public Node From { get; set; }
         public Node To { get; set; }
-        public double DistanceInMeters { get; set; }
-        public double DurationInSeconds { get; set; }
+        public int DistanceInMeters { get; set; }
+        public int DurationInSeconds { get; set; }
         public double Costs { get; set; }
 
         public static Distance CreateFromDTO(DistanceDTO distance, Node from, Node to)
@@ -25,20 +24,20 @@ namespace ResearchPlatform.Models
             };
         }
 
-        private static double CalculateAvg(List<List<int>> distances)
+        private static int CalculateAvg(List<List<int>> distances)
         {
             var summedRows = distances
                     .Select(row => row.Sum())
                     .ToList();
-            return summedRows.Average();
+            return (int) summedRows.Average();
         }
 
-        private static double CalculateAvg(List<List<double>> distances)
+        private static int CalculateAvg(List<List<double>> distances)
         {
             var summedRows = distances
                     .Select(row => row.Sum())
                     .ToList();
-            return summedRows.Average();
+            return (int) summedRows.Average();
         }
     }
